requirements.txt
boxmot==10.0.51
numpy==1.23.5
opencv_python==4.8.0.74
opencv_python_headless==4.7.0.72
Pillow==9.4.0
Pillow==10.2.0
supervision==0.18.0
torch==2.0.1
torchvision==0.15.2
ultralytics==8.1.14
